requirements.txt
discord.py
openai
difflib
aiohttp
pynacl
praw
beautifulsoup4